+++ firebase-database/src/main/java/com/google/firebase/database/core/SyncTree.java
@@ -851,16 +851,18 @@ public class SyncTree {
     keepSynced(query, keep, false);
   }
 
-  public void keepSynced(final QuerySpec query, final boolean keep, final boolean skipDedup) {
+  public boolean keepSynced(final QuerySpec query, final boolean keep, final boolean skipDedup) {
     if (keep && !keepSyncedQueries.contains(query)) {
       // TODO[persistence]: Find better / more efficient way to do keep-synced listeners.
       addEventRegistration(
           new KeepSyncedEventRegistration(query), /* skipListenerSetup= */ skipDedup);
       keepSyncedQueries.add(query);
+      return true;
     } else if (!keep && keepSyncedQueries.contains(query)) {
       removeEventRegistration(new KeepSyncedEventRegistration(query), skipDedup);
       keepSyncedQueries.remove(query);
     }
+    return false;
   }
 
   /**