+++ firebase-database/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [fixed] Fixed a crash when calling `keepSynced(true)` and `get()` on the same query.
+
 # 22.0.1
 
 - [changed] Bumped internal dependencies.